[kotlin-migrate] Migrate BaseActivity, MainActivity and FacilityPage to kotlin
These 3 activities have to be migrated together so that we don't have to deal with slightly different override enforcement in Java and kotlin.
When it's migrated to Kotlin, we can use a much more succint function type syntax `(Boolean) -> Unit`. Therefore, I was able to get rid of the `return null` lines in callback just to make Java happy.