Setup CI workflows (#12)
Summary:
This commit aims to setup CI workflows for this repository so that we can confidently migrating to Kotlin and add new features later.
This is only a starting point. Linter (and ideally tests) should be added later.

Test Plan:
You can directly see the build status in the pull request page once the pull request is marked ready for review.